remove a few gcc 4 warnings   work around freebsd 5.4 brokenness (if you don't have IPv6 in the     kernel, socket(PF_INET6,SOCK_STREAM,0) returns EPROTONOSUPPORT     instead of EPFNOSUPPORT, which basically says "yeah, I know IPv6,     but TCP?  never heard of it")

+++ dns/dns_rcip.c
@@ -10,8 +10,8 @@ static stralloc data = {0};
 
 static int init(char ip[256])
 {
-  int i;
-  int j;
+  unsigned long int i;
+  unsigned long int j;
   int iplen = 0;
   char *x;
 
@@ -30,7 +30,7 @@ static int init(char ip[256])
 
   if (!iplen) {
     i = openreadclose("/etc/resolv.conf",&data,64);
-    if (i == -1) return -1;
+    if (i == (unsigned long int)-1) return -1;
     if (i) {
       if (!stralloc_append(&data,"\n")) return -1;
       i = 0;
